requirements.txt
streamlit
langchain
langchain-openai
langchain-community
openai
python-dotenv
youtube-transcript-api
faiss-cpu
langchain_google_genai
